requirements.txt
certifi==2019.9.11
python-dotenv